# Implement some minor performance optimizations