--- a/compiler/parser/Cargo.toml
+++ b/compiler/parser/Cargo.toml
@@ -32,6 +32,7 @@ unic-emoji-char = "0.9.0"
 unic-ucd-ident  = "0.9.0"
 unicode_names2 = "0.5.0"
 thiserror = "1.0"
+fnv = "1.0.7"
 
 [dev-dependencies]
 insta = "1.14.0"