--- a/compiler/parser/Cargo.toml
+++ b/compiler/parser/Cargo.toml
@@ -28,11 +28,11 @@ log = "0.4.16"
 num-bigint = "0.4.3"
 num-traits = "0.2.14"
 phf = "0.10.1"
+rustc-hash = "1.1.0"
+thiserror = "1.0"
 unic-emoji-char = "0.9.0"
 unic-ucd-ident  = "0.9.0"
 unicode_names2 = "0.5.0"
-thiserror = "1.0"
-fnv = "1.0.7"
 
 [dev-dependencies]
 insta = "1.14.0"