I tend to benchmark the parser by parsing all of CPython in [Ruff](https://github.com/charliermarsh/ruff). This involves (1) disabling Ruff's linting pass, then (2) running `cargo build --release && hyperfine --warmup 10 --runs 50 -i "./target/release/ruff resources/test/cpython --no-cache"`.

Prior to this change:

```
Benchmark 1: ./target/release/ruff resources/test/cpython --no-cache
  Time (mean ± σ):     265.2 ms ±   4.6 ms    [User: 2260.7 ms, System: 83.1 ms]
  Range (min … max):   259.0 ms … 281.1 ms    50 runs

Benchmark 1: ./target/release/ruff resources/test/cpython --no-cache
  Time (mean ± σ):     268.4 ms ±   4.2 ms    [User: 2268.1 ms, System: 83.5 ms]
  Range (min … max):   262.5 ms … 283.6 ms    50 runs

Benchmark 1: ./target/release/ruff resources/test/cpython --no-cache
  Time (mean ± σ):     269.7 ms ±   5.5 ms    [User: 2277.9 ms, System: 83.5 ms]
  Range (min … max):   260.1 ms … 280.7 ms    50 runs
```

After this change:

```
Benchmark 1: ./target/release/ruff resources/test/cpython --no-cache
  Time (mean ± σ):     255.6 ms ±   5.1 ms    [User: 2156.0 ms, System: 83.4 ms]
  Range (min … max):   247.2 ms … 271.8 ms    50 runs

Benchmark 1: ./target/release/ruff resources/test/cpython --no-cache
  Time (mean ± σ):     255.4 ms ±   4.8 ms    [User: 2152.8 ms, System: 83.0 ms]
  Range (min … max):   246.9 ms … 269.7 ms    50 runs

Benchmark 1: ./target/release/ruff resources/test/cpython --no-cache
  Time (mean ± σ):     256.3 ms ±   4.8 ms    [User: 2156.3 ms, System: 82.8 ms]
  Range (min … max):   248.5 ms … 270.6 ms    50 runs
```

So it's a ~4.8% improvement, if you take the median results.

Note that there's about ~58ms of overhead associated with this benchmark (e.g., to locate all the Python files on-disk, read their contents, etc.), so that's really a lower-bound on improvement.