@charliermarsh 2022-11-21
@coolreader18 - Definitely! We've used it from the very beginnings of Ruff. I've tried to contribute back as many compatibility improvements as I can :)